extract a common metadata layer from annotationSet component
