+++ projects/media-viewer/src/assets/sass/annotation-set.scss
@@ -1,4 +1,4 @@
-mv-annotation-set, mv-redactions, mv-bookmark-icons {
+mv-metadata-layer, mv-redactions, mv-bookmark-icons {
   .rotation {
     -webkit-transform-origin: left bottom;
     -moz-transform-origin: left bottom;
