+++ projects/media-viewer/src/assets/sass/pdf-viewer.scss
@@ -18,7 +18,7 @@ mv-pdf-viewer {
     z-index: 1000;
   }
 
-  .drawMode > .page > mv-annotation-set > div {
+  .drawMode > .page > mv-metadata-layer > div {
     z-index: 1000;
   }
