+++ projects/media-viewer/src/lib/viewers/pdf-viewer/pdf-viewer.component.html
@@ -23,6 +23,7 @@
   <div grabNDrag [dragX]="enableGrabNDrag"
        #pdfViewer
        class="pdfViewer"
+       mvCreateTextHighlight
        [ngClass]="{ hidden: loadingDocument, highlightMode: highlightMode | async, drawMode: drawMode | async }">
   </div>
 
@@ -32,11 +33,11 @@
   ></mv-redactions>
 
   <ng-template #annotationTemplate>
-    <mv-annotation-set
+    <mv-metadata-layer
       *ngIf="enableAnnotations && annotationSet"
       [zoom]="zoom"
       [rotate]="rotation">
-    </mv-annotation-set>
+    </mv-metadata-layer>
     <mv-comment-set #commentsPanel
                     *ngIf="enableAnnotations && annotationSet"
                     [annotationSet]="annotationSet"
